Sketch of a pom.xml for the main line repository
This is a sketch- it isn't ready to merge into master.
We have a number of customers who are interested in using Maven
in their builds, so having an example pom.xml as a convenience for adding the
Mixpanel android api into a local or organization level repository
would be useful.

However, Maven is not our canonical build tool, and we don't want
to actively support and maintain multiple builds.
That means it is much more likely we can merge a pom.xml if it
delegates all of the work of compiling and building to our
current canonical build, which is ant.

The proposed pom.xml in this commit demonstrates, in broad strokes,
the sort of pom.xml we could consider for inclusion into the main
line of development. Unfortunately, it isn't complete (for example,
it doesn't identify the canonical build product, MixpanelAPI.jar
in the root directory.)